task07.task0724 - Family census
Create the Human class with String name, boolean sex, int age, Human father, and Human mother fields. Create and populate objects so that we end up with: Two grandfathers, two grandmothers, one father, one mother, and three children. Display the objects on the screen.

Note:
If you write your own String toString() method in the Human class, then it will be used to display the object on the screen.