Create task07.task0727 - Changing functionalit
Task: The program takes in strings until the user enters an empty string (by pressing Enter). Then it converts the strings to uppercase ("mother" becomes "MOTHER") and displays them.

New task: The program takes in strings until the user enters an empty string (by pressing Enter).
Then the program builds a new list. If a string has an even number of letters, then the string is duplicated. If the number of letters is odd, then the string is repeated three times.
The program displays the contents of the new list.